requirements.txt
absl-py==0.1.10
annoy==1.11.5
appnope==0.1.0
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
#certifi==2018.1.18
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
constantly==15.1.0
cycler==0.10.0
Cython==0.27.3
dask==0.17.1
decorator==4.2.1
enum34==1.1.6
Flask==0.12.2
funcsigs==1.0.2
futures==3.2.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.6
ImageHash==4.0
#imageio==2.2.0
incremental==17.5.0
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
line-profiler==2.1.2
locket==0.2.0
lshash==0.0.4.dev0
mahotas==1.4.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mido==1.2.8
mock==2.0.0
networkx==2.1
nose==1.3.7
numexpr==2.6.4
numpy==1.14.2
pandas==0.22.0
partd==0.3.8
pathlib2==2.3.0
pbr==3.1.1
peewee==3.0.10
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.4.3
#ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-magic==0.4.15
python-rtmidi==1.1.0
pytz==2018.3
PyWavelets==0.5.2
requests==2.18.4
scandir==1.6
-e git+https://github.com/scikit-image/scikit-image.git@db80d0397875964e8c3de579a3951de924b10959#egg=scikit_image
scikit-learn==0.19.1
scipy==1.0.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
tables==3.4.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
toolz==0.9.0
tqdm==4.19.5
traitlets==4.3.2
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
uWSGI==2.0.16
wcwidth==0.1.7
Werkzeug==0.14.1
zope.interface==4.4.3
tweepy==3.6.0

